Remove non-functional restart keymap
The restart functionality was attempting to pass file arguments to the :restart command,
which is not supported in Neovim. The :restart command doesn't accept arguments and
always restarts fresh without preserving the current file context.

Removed the <leader>rs keymap and related code that attempted to use vim.cmd.restart(file)
since this feature was non-functional and causing more trouble than it provided.